:bug: Fix namings api
