+++ docs/marketing_automation/introduction.mdx
@@ -1,6 +1,6 @@
 ---
-title: 'Coming Soon'
+title: "Coming Soon"
 description: "You're a bit early!"
 ---
 
-## [<u>Click here</u>](https://tally.so/r/3Exrdl?integration_category=marketing_automation) to be notified once this feature is live!
+## [<u>Click here</u>](https://tally.so/r/3Exrdl?integration_category=marketingautomation) to be notified once this feature is live!
