+++ docs/mint.json
@@ -56,7 +56,7 @@
     },
     {
       "name": "Marketing Automation",
-      "url": "marketing_automation"
+      "url": "marketingautomation"
     },
     {
       "name": "Ticketing",
@@ -116,7 +116,7 @@
     {
       "group": "Marketing Automation",
       "pages": [
-        "marketing_automation/introduction"
+        "marketingautomation/introduction"
       ]
     },
     {
