+++ docs/open-source/contributors.mdx
@@ -1,7 +1,7 @@
 ---
-title: 'Build a connector'
-description: 'We welcome all contributions to Panora; from small UI enhancements to brand new integrations. We love seeing community members level up and give people power-ups!'
-icon: 'star'
+title: "Build a connector"
+description: "We welcome all contributions to Panora; from small UI enhancements to brand new integrations. We love seeing community members level up and give people power-ups!"
+icon: "star"
 ---
 
 ## Introduction
@@ -14,9 +14,7 @@ We made a docker file that builds Panora from sources, specifically to help you 
 
 <Steps>
 
-
 <Step title="Copy env variables">```bash cp .env.example .env ```</Step>
-
 
 <Step title="Removed previously installed dependencies">
   ```bash rm -rf node_modules .pnpm-store ./packages/api/dist
@@ -53,9 +51,9 @@ First choose wisely which vertical the 3rd party belongs to among these:
 - ticketing
 - accounting
 - ats
-- file_storage
+- filestorage
 - hris
-- marketing_automation
+- marketingautomation
 
 For the sake of the guide, now on we'll consider adding a 3rd party belonging to the `crm` vertical.
