+++ packages/api/src/@core/connections/marketingautomation/marketingautomation.connection.module.ts
@@ -6,7 +6,7 @@ import { WebhookModule } from '@@core/webhook/webhook.module';
 import { EnvironmentService } from '@@core/environment/environment.service';
 import { EncryptionService } from '@@core/encryption/encryption.service';
 import { ConnectionsStrategiesService } from '@@core/connections-strategies/connections-strategies.service';
-import { MarketingAutomationConnectionsService } from './services/marketing_automation.connection.service';
+import { MarketingAutomationConnectionsService } from './services/marketingautomation.connection.service';
 import { ServiceRegistry } from './services/registry.service';
 
 @Module({
