+++ packages/api/src/@core/connections/marketingautomation/services/marketingautomation.connection.service.ts
