+++ packages/api/src/@core/connections/marketingautomation/types/index.ts
