+++ packages/api/src/app.module.ts
@@ -7,10 +7,10 @@ import { ScheduleModule } from '@nestjs/schedule';
 import { TasksService } from './@core/tasks/tasks.service';
 import { LoggerModule } from 'nestjs-pino';
 import { HrisModule } from './hris/hris.module';
-import { MarketingAutomationModule } from './marketing_automation/marketing-automation.module';
+import { MarketingAutomationModule } from './marketingautomation/marketingautomation.module';
 import { AtsModule } from './ats/ats.module';
 import { AccountingModule } from './accounting/accounting.module';
-import { FileStorageModule } from './file_storage/file-storage.module';
+import { FileStorageModule } from './filestorage/filestorage.module';
 import { SentryInterceptor, SentryModule } from '@ntegral/nestjs-sentry';
 import { APP_GUARD, APP_INTERCEPTOR } from '@nestjs/core';
 import { LoggerService } from '@@core/logger/logger.service';
