+++ packages/api/src/marketingautomation/marketingautomation.module.ts
