+++ packages/api/tsconfig.json
@@ -16,8 +16,8 @@
       "@crm/*": ["src/crm/*"],
       "@hris/*": ["src/hris/*"],
       "@ticketing/*": ["src/ticketing/*"],
-      "@file-storage/*": ["src/file_storage/*"],
-      "@marketing-automation/*": ["src/marketing_automation/*"],
+      "@filestorage/*": ["src/filestorage/*"],
+      "@marketingautomation/*": ["src/marketingautomation/*"],
       "@ats/*": ["src/ats/*"],
       "@accounting/*": ["src/accounting/*"],
     },
