+++ packages/shared/src/providers.ts
@@ -8,8 +8,8 @@ export const HRIS_PROVIDERS = [''];
 export const ATS_PROVIDERS = [''];
 export const ACCOUNTING_PROVIDERS = [''];
 export const TICKETING_PROVIDERS = ['zendesk', 'front', 'github', 'jira', 'gorgias'];
-export const MARKETING_AUTOMATION_PROVIDERS = [''];
-export const FILE_STORAGE_PROVIDERS = [''];
+export const MARKETINGAUTOMATION_PROVIDERS = [''];
+export const FILESTORAGE_PROVIDERS = [''];
  
 
 export function getProviderVertical(providerName: string): ProviderVertical {
@@ -28,10 +28,10 @@ export function getProviderVertical(providerName: string): ProviderVertical {
   if (TICKETING_PROVIDERS.includes(providerName)) {
     return ProviderVertical.Ticketing;
   }
-  if (MARKETING_AUTOMATION_PROVIDERS.includes(providerName)) {
+  if (MARKETINGAUTOMATION_PROVIDERS.includes(providerName)) {
     return ProviderVertical.MarketingAutomation;
   }
-  if (FILE_STORAGE_PROVIDERS.includes(providerName)) {
+  if (FILESTORAGE_PROVIDERS.includes(providerName)) {
     return ProviderVertical.FileStorage;
   }
   return ProviderVertical.Unknown;
