+++ packages/shared/src/utils.ts
@@ -4,6 +4,10 @@ export enum AuthStrategy {
   basic
 }
 
+export enum SoftwareMode {
+  cloud = "CLOUD",
+}
+
 export type ProviderConfig = {
   scopes: string;
   logoPath: string;
@@ -876,7 +880,7 @@ export const providersConfig: ProvidersConfig = {
       authStrategy: AuthStrategy.oauth2
     },
   },
-  'marketing_automation': {
+  'marketingautomation': {
     'active_campaign': {
       scopes: '',
       urls: {
